Fix `cls.syntax` deprecation warning and use `args`
In Sublime Text 3, there is was a `cls.syntax` deprecation warning in
the console as a result of using:

```python
syntax = 'ruby haml'
```

This removes this warning by using the `selector` configuration:

```python
defaults = {
    'selector': 'text.haml'
}
```

On top of this, it is no longer required to have `.haml-lint.yml`
present and `haml_lint` can make use of the default linting rules.

The `args` and `temp_file` to lint can be specified in the
`SublimeLinter.sublime-settings` file.